% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetacellToolkit-package.R
\docType{data}
\name{CD34_mc}
\alias{CD34_mc}
\title{CD34 dataset at the metacell level}
\format{
A Seurat object containing the metacells data, *i.e.* gene expression matrix as well as the membership of each single-cell to a metacell in the *misc* slot of the Seurat object.
}
\source{
\url{https://www.nature.com/articles/s41587-023-01716-9}
}
\usage{
CD34_mc
}
\description{
Metacells derived from a single-cell dataset of CD34 cells obtained from \href{https://www.nature.com/articles/s41587-023-01716-9}{Persad et al. (2023)} and available 
at the single-cell level in MetacellToolkit::CD34_sc.
Metacells data were generated using the command lines based on SuperCell and described in \href{https://github.com/GfellerLab/MetacellToolkit}{the MetacellToolkit github repository}.
}
\keyword{datasets}
