% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetacellToolkit-package.R
\docType{data}
\name{CD34_sc}
\alias{CD34_sc}
\title{CD34 dataset at the single-cell level}
\format{
An Seurat object containing the single-cell data (raw counts and metadata).
}
\source{
\url{https://www.nature.com/articles/s41587-023-01716-9}
}
\usage{
CD34_sc
}
\description{
Single-cell dataset composed of 6881 CD34 cells obtained from \href{https://www.nature.com/articles/s41587-023-01716-9}{Persad et al. (2023)}.
}
\keyword{datasets}
