% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diffusion_comp.R
\name{get_diffusion_comp}
\alias{get_diffusion_comp}
\title{Compute separation of metacells.}
\usage{
get_diffusion_comp(sc.obj, sc.reduction = "pca", dims = NULL, n.diff.comp = 10)
}
\arguments{
\item{sc.obj}{A Seurat object containing the single-cell data from which the metacells were built.}

\item{sc.reduction}{(optional, default is "pca") A string indicating which low embedding from sc.obj should be used to compute the diffusion 
components or a data frame containing a pre-computed embedding of the single-cell data.}

\item{dims}{(optional, default is NULL) Vector indicating the embedding components that should be used to compute the compactness.}

\item{n.diff.comp}{(optional, default is 10) Number of eigen vectors to use in the determine_multiscale_space function in palantir.}
}
\value{
A data.frame containing the diffusion components.
}
\description{
\code{get_diffusion_comp} 
This function computes diffusion components from palantir starting from a Seurat object (relies on python functions adapted from https://github.com/dpeerlab/SEACells/blob/main/SEACells/evaluate.py).
}
\examples{
get_diffusion_comp(sc.obj = CD34_sc, dims = 1:30)
}
