% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dim_reduc.R
\name{get_dim_reduc}
\alias{get_dim_reduc}
\title{Compute separation of metacells.}
\usage{
get_dim_reduc(sc.obj, sc.reduction = "pca", n.components = 10)
}
\arguments{
\item{sc.obj}{A Seurat object containing the single-cell data from which the metacells were built.}

\item{sc.reduction}{(optional, default is "pca") A string indicating which low embedding from sc.obj should be used to compute the diffusion 
components or a data frame containing a pre-computed embedding of the single-cell data.}

\item{n.components}{(optional, default is 10) Number of embedding components that should be used to compute diffusion maps.}
}
\value{
A data.frame containing the diffusion components.
}
\description{
\code{get_dim_reduc} 
This function computes diffusion components from palantir starting from a Seurat object (relies on python functions adapted from https://github.com/dpeerlab/SEACells/blob/main/SEACells/evaluate.py).
}
\examples{
get_dim_reduc(sc.obj = CD34_sc, n.components = 30)
}
