% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_compactness.R
\name{mc_compactness}
\alias{mc_compactness}
\title{Compute compactness of metacells.}
\usage{
mc_compactness(
  cell.membership,
  sc.obj,
  sc.reduction,
  group.label = "membership",
  dims = NULL
)
}
\arguments{
\item{cell.membership}{A data frame containing at least one column assigning each single-cell to a metacell (named "membership" if no other label is provided in group.label)
and single-cell IDs as rownames.}

\item{sc.obj}{A Seurat object containing the single-cell data from which the metacells were built.}

\item{sc.reduction}{(optional, default is "pca") A string indicating which low embedding from sc.obj should be used to compute compactness or
a data frame containing a pre-computed embedding of the single-cell data.}

\item{group.label}{(optional, default is "membership") A string indicating the column name from *cell.membership* that should be used to compute 
the compactness metric.}

\item{dims}{(optional, default is NULL) Vector indicating the embedding components that should be used to compute the compactness. 
If NULL, components 1 to max number of components in the embedded space will be used.}
}
\value{
A vector containing the compactness of each metacell.
}
\description{
\code{mc_compactness} 
This function computes the compactness of metacells (relies on python functions adapted from https://github.com/dpeerlab/SEACells/blob/main/SEACells/evaluate.py).
}
\examples{
mc_compactness(cell.membership = CD34_mc@misc$cell_membership$membership, sc.obj = CD34_sc, sc.reduction = "pca")
}
