% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_INV.R
\name{mc_INV}
\alias{mc_INV}
\title{Compute inner-normalized variance (INV) of metacells.}
\usage{
mc_INV(
  sc.obj,
  cell.membership,
  group.label = "membership",
  assay = "RNA",
  slot = "counts",
  do.norm = T,
  NA_val = 1,
  scaling_factor = NULL
)
}
\arguments{
\item{sc.obj}{A Seurat object containing the single-cell data from which the metacells were built.}

\item{cell.membership}{A data frame containing at least one column assigning each single-cell to a metacell (named "membership" if no other label is provided in group.label)
and single-cell IDs as rownames.}

\item{group.label}{(optional, default is "membership") A string indicating the column name from *cell.membership* that should be used to compute the INV metric.}

\item{assay}{(optional, default is "RNA") A string indicating the assay from *sc.obj* that should be used to compute the INV metric.}

\item{slot}{(optional, default is "counts") A string indicating the slot/layer from *sc.obj* that should be used to compute the INV metric.}

\item{do.norm}{(optional, default is T) Logical value indicating whether the input slot data should be normalized (normalization performed within each metacell).}

\item{NA_val}{(optional, default is 1) Numeric value indicating which value should be used to replace NA values.}

\item{scaling_factor}{(optional, default is NULL) Used only if *do.norm* is TRUE. Numeric value indicating which scaling factor should be used in the counts normalization. 
If NULL, the median of the total number of UMIs in the cells belonging to the metacell is used.}
}
\value{
A vector containing the INV value of each metacell.
}
\description{
\code{mc_INV} 
This function computes the inner-normalized variance (INV) of metacells (relies on functions inspired from https://github.com/tanaylab/metacells/blob/master/metacells/tools/quality.py).
}
\examples{
mc_INV(sc.obj = MetacellAnalysisToolkit::CD34_sc, cell.membership = MetacellAnalysisToolkit::CD34_mc@misc$cell_membership)
}
