% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_projection.R
\name{mc_projection}
\alias{mc_projection}
\title{mc_projection of metacells in the single cell space}
\usage{
mc_projection(
  sc.obj,
  mc.obj,
  membership = NULL,
  metacell.label = NULL,
  sc.label = NULL,
  dims = c(1, 2),
  sc.reduction = "umap",
  mc.color = NULL,
  sc.color = NULL,
  alpha = 1,
  pt_size = 0,
  metric = "size",
  continuous_metric = F,
  seed = 1234
)
}
\arguments{
\item{sc.obj}{A Seurat object containing the single-cell data from which the metacells were built.}

\item{mc.obj}{A Seurat object containing the metacells data.}

\item{membership}{(optional, default is NULL) A vector containing the assignment of each single-cell to a metacell. 
If *membership* is NULL, the membership information will be retrieved from the *misc* slot in the *mc.obj* under the assumption that the metacell Seurat object
was generated using the MetacellToolkit command lines (see format of the *MetacellToolkit::CD34_mc* object).}

\item{metacell.label}{String corresponding to the name of the metadata column in mc.obj that should be used to color metacells.}

\item{sc.label}{String corresponding to the name of the metadata column in sc.obj that should be used to color single cells.}

\item{dims}{Numerical vector (of length 2) indicating which components should be used in the 2D visualization of the data.}

\item{sc.reduction}{(optional, default is "umap") A string indicating which low embedding from sc.obj should be used for the 2D visualization of the data
or a data frame containing a pre-computed embedding of the single-cell data. If *sc.reduction* is a string which does not correspond to any embedding from sc.obj
PCA will be performed on the single-cell data and UMAP will be run to obtain the 2D representation.}

\item{mc.color}{Colors for metacell idents}

\item{sc.color}{Colors for single-cell idents}

\item{alpha}{Transparency value for the single-cell points}

\item{pt_size}{Size the single-cell points}

\item{metric}{Column name of a continuous metric to use either to color the metacells (if continuous_metric is TRUE) or to define metacells sizes (if continuous_metric is TRUE).}

\item{continuous_metric}{Bolean indicating if the metric variable is continuous or not. If TRUE a continuous color scale will be used for the metacell colors.}
}
\value{
A plot of metacells projected in the single cell space.
}
\description{
\code{mc_projection} 
This function plots metacells in a single-cell space (pca, umap,..) taking the average coordinates of single cells in each metacell
}
\examples{
mc_projection(sc.obj = CD34_sc, mc.obj = CD34_mc)
mc_projection(sc.obj = CD34_sc, mc.obj = CD34_mc, metric = "celltype_purity", continuous_metric = T)
mc_projection(sc.obj = CD34_sc, mc.obj = CD34_mc, metric = "celltype_purity", continuous_metric = T, sc.label = "celltype")
mc_projection(sc.obj = CD34_sc, mc.obj = CD34_mc, sc.label = "celltype", metacell.label = "celltype")

}
