% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_purity.R
\name{mc_purity}
\alias{mc_purity}
\title{Compute purity of metacells.}
\usage{
mc_purity(membership, annotation, method = c("max_proportion", "entropy")[1])
}
\arguments{
\item{membership:}{Membership vector associating each single-cell to a metacell.}

\item{annotation:}{Vector of single-cell annotations of length equal to the number of single-cells used 
to build the metacell object. The annotations order should match the single-cell membership provided in the membership parameter.}

\item{method:}{Method to compute metacell purity. "max_proportion" if the purity is defined as a proportion of the most abundant annotation group (e.g. cell type) 
within the metacell or "entropy" if the purity is defined as the Shanon entropy of the annotation groups metacells consists of.}
}
\value{
A vector containing the purity of each metacell.
}
\description{
\code{mc_purity} 
This function computes the purity of metacells. (code has been retrieved from the SuperCell::supercell_purity function).
}
\examples{
mc_purity(membership = CD34_mc@misc$cell_membership$membership, annotation = CD34_sc[['celltype']])
}
