% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_separation.R
\name{mc_separation}
\alias{mc_separation}
\title{Compute separation of metacells.}
\usage{
mc_separation(
  cell.membership,
  sc.obj = NULL,
  sc.reduction = "pca",
  group.label = "membership",
  dims = NULL,
  nth.nbr = 1
)
}
\arguments{
\item{cell.membership}{A data frame containing at least one column assigning each single-cell to a metacell (named "membership" if no other label is provided in group.label)
and single-cell IDs as rownames.}

\item{sc.obj}{A Seurat object containing the single-cell data from which the metacells were built.}

\item{sc.reduction}{(optional, default is "pca") A string indicating which low embedding from sc.obj should be used to compute compactness or
a data frame containing a pre-computed embedding of the single-cell data.}

\item{group.label}{(optional, default is "membership") A string indicating the column name from *cell.membership* that should be used to compute 
the compactness metric.}

\item{dims}{(optional, default is NULL) Vector indicating the embedding components that should be used to compute the compactness. 
If NULL, components 1 to max number of components in the embedded space will be used.}

\item{nth.nbr}{(optional, default is 1) Number of nearest neighbors used to compute the separation metric.}
}
\value{
A vector containing the separation metric for each metacell.
}
\description{
\code{mc_separation} 
This function computes the separation of metacells (relies on python functions adapted from https://github.com/dpeerlab/SEACells/blob/main/SEACells/evaluate.py).
}
\examples{
mc_separation(cell.membership = CD34_mc@misc$cell_membership, sc.obj = CD34_sc)
}
